restore sample mapping file and refactor save helpers
