+++ client/src/TargetsAgent.js
@@ -1,20 +1,9 @@
+import { groupByLayer, removeLayerEntries } from './utils/entryHelpers.js';
+
 export function groupTargetsByLayer(data) {
-  const result = {};
-  if (!data.Targets) return result;
-  Object.entries(data.Targets).forEach(([key, value]) => {
-    const [layer, indexPart] = key.split('.');
-    const decIndex = parseInt(indexPart, 10);
-    const hexVal = parseInt(value, 16);
-    const offset = decIndex - hexVal;
-    if (!result[layer]) result[layer] = [];
-    result[layer].push({ key, value, offset });
-  });
-  return result;
+  return groupByLayer(data.Targets || {});
 }
 
 export function removeLayerTargets(data, layer) {
-  if (!data.Targets) return;
-  Object.keys(data.Targets).forEach(key => {
-    if (key.startsWith(`${layer}.`)) delete data.Targets[key];
-  });
+  if (data.Targets) removeLayerEntries(data.Targets, layer);
 }
