+++ client/src/hooks/useMappingEditor.js
@@ -11,6 +11,7 @@ import {
   groupSourcesByLayer,
   removeLayerSources,
 } from '../SourcesAgent.js';
+import { setLayerEntries } from '../utils/entryHelpers.js';
 
 export default function useMappingEditor() {
   const [iniData, setIniData] = useState(null);
@@ -126,24 +127,14 @@ export default function useMappingEditor() {
 
   const saveTargets = useCallback((layerKey, entries) => {
     const dataCopy = { ...iniData, Targets: { ...iniData.Targets } };
-    Object.keys(dataCopy.Targets).forEach(k => {
-      if (k.startsWith(`${layerKey}.`)) delete dataCopy.Targets[k];
-    });
-    entries.forEach(e => {
-      dataCopy.Targets[e.key] = e.value;
-    });
+    setLayerEntries(dataCopy.Targets, layerKey, entries);
     setIniData(dataCopy);
     setTargets(groupTargetsByLayer(dataCopy));
   }, [iniData]);
 
   const saveSources = useCallback((layerKey, entries) => {
     const dataCopy = { ...iniData, Sources: { ...iniData.Sources } };
-    Object.keys(dataCopy.Sources).forEach(k => {
-      if (k.startsWith(`${layerKey}.`)) delete dataCopy.Sources[k];
-    });
-    entries.forEach(e => {
-      dataCopy.Sources[e.key] = e.value;
-    });
+    setLayerEntries(dataCopy.Sources, layerKey, entries);
     setIniData(dataCopy);
     setSources(groupSourcesByLayer(dataCopy));
   }, [iniData]);
